Title: Tanzanian Water Pumps Problem (Mar' 21) 1. Performed data cleansing, pre-processing and feature engineering on Tanzanian Water pumps dataset ,using SVM, XG Boost, Random Forest and Decision Trees classification techniques. 2. Diligent hyper-parameter tuning was performed for Decision Tree and Random Forests models
The ministry of water of an African country has made a data set of pumps installed in various places in the country to serve the water for their countrymen. It is a tedious and expensive task to maintain these pumps. This data set contains information such as the kind of pump, when it was installed, and how it is managed. Can you predict which pumps require repairs and which are not functional by using the given data set? A smart understanding of water point failure can improve maintenance operations and ensure that clean and safe water is available to these communities. Experiment with a few methods such as tree based methods, bagging, boosting method, support vector machine etc. and comment on the suitability of the method to predict the classes. Before implementing any model, visualize the data and comment on the data statistics.